+++ Realm/Tests/TestHost/main.m
@@ -8,7 +8,13 @@
 
 #import <TargetConditionals.h>
 
-#if TARGET_OS_IPHONE
+#if TARGET_OS_TV || TARGET_OS_WATCH
+
+// tvOS and watchOS don't support testing at this time.
+int main(int argc, const char *argv[]) {
+}
+
+#elif TARGET_OS_IPHONE
 
 #import <UIKit/UIKit.h>