+++ build.sh
@@ -101,14 +101,8 @@ xc() {
     fi
 }
 
-xcrealm() {
-    PROJECT=Realm.xcodeproj
-    xc "-project $PROJECT $@"
-}
-
-xcrealmswift() {
-    PROJECT=RealmSwift.xcodeproj
-    xc "-project $PROJECT $@"
+copy_bcsymbolmap() {
+    find "$1" -name '*.bcsymbolmap' -type f -exec cp {} "$2" \;
 }
 
 build_ios_combined() {
@@ -119,7 +113,7 @@ build_ios_combined() {
     local config="$CONFIGURATION"
 
     # Derive build paths
-    local build_products_path="build/DerivedData/$module_name/Build/Products"
+    local build_products_path="build/DerivedData/Realm/Build/Products"
     local product_name="$module_name.framework"
     local binary_path="$module_name"
     local iphoneos_path="$build_products_path/$config-iphoneos$scope_suffix/$product_name"
@@ -127,14 +121,16 @@ build_ios_combined() {
     local out_path="build/ios$scope_suffix$version_suffix"
 
     # Build for each platform
-    cmd=$(echo "xc$module_name" | tr '[:upper:]' '[:lower:]') # lowercase the module name to generate command (xcrealm or xcrealmswift)
-    $cmd "-scheme '$scheme' -configuration $config -sdk iphoneos"
-    $cmd "-scheme '$scheme' -configuration $config -sdk iphonesimulator ONLY_ACTIVE_ARCH=NO"
+    xc "-scheme '$scheme' -configuration $config -sdk iphoneos"
+    xc "-scheme '$scheme' -configuration $config -sdk iphonesimulator -destination 'name=iPhone 6' ONLY_ACTIVE_ARCH=NO"
 
     # Combine .swiftmodule
     if [ -d $iphonesimulator_path/Modules/$module_name.swiftmodule ]; then
       cp $iphonesimulator_path/Modules/$module_name.swiftmodule/* $iphoneos_path/Modules/$module_name.swiftmodule/
     fi
+    
+    # Copy *.bcsymbolmap to .framework for submitting app with bitcode
+    copy_bcsymbolmap "$build_products_path/$config-iphoneos$scope_suffix" "$iphoneos_path"
 
     # Retrieve build products
     clean_retrieve $iphoneos_path $out_path $product_name
@@ -150,7 +146,7 @@ build_watchos_combined() {
     local config="$CONFIGURATION"
 
     # Derive build paths
-    local build_products_path="build/DerivedData/$module_name/Build/Products"
+    local build_products_path="build/DerivedData/Realm/Build/Products"
     local product_name="$module_name.framework"
     local binary_path="$module_name"
     local watchos_path="$build_products_path/$config-watchos$scope_suffix/$product_name"
@@ -158,20 +154,30 @@ build_watchos_combined() {
     local out_path="build/watchos$scope_suffix"
 
     # Build for each platform
-    cmd=$(echo "xc$module_name" | tr '[:upper:]' '[:lower:]') # lowercase the module name to generate command (xcrealm or xcrealmswift)
-    $cmd "-scheme '$scheme' -configuration $config -sdk watchos"
-    $cmd "-scheme '$scheme' -configuration $config -sdk watchsimulator ONLY_ACTIVE_ARCH=NO"
+    xc "-scheme '$scheme' -configuration $config -sdk watchos"
+    xc "-scheme '$scheme' -configuration $config -sdk watchsimulator -destination 'name=Apple Watch - 42mm' ONLY_ACTIVE_ARCH=NO"
 
     # Combine .swiftmodule
     if [ -d $watchsimulator_path/Modules/$module_name.swiftmodule ]; then
       cp $watchsimulator_path/Modules/$module_name.swiftmodule/* $watchos_path/Modules/$module_name.swiftmodule/
     fi
+    
+    # Copy *.bcsymbolmap
+    copy_bcsymbolmap "$build_products_path/$config-watchos$scope_suffix" "$watchos_path"
 
     # Retrieve build products
     clean_retrieve $watchos_path $out_path $product_name
 
     # Combine ar archives
     xcrun lipo -create "$watchsimulator_path/$binary_path" "$watchos_path/$binary_path" -output "$out_path/$product_name/$module_name"
+}
+
+xc_work_around_rdar_23055637() {
+    # xcodebuild times out waiting for the iOS simulator to launch if it takes > 120 seconds for the tests to
+    # build (<http://openradar.appspot.com/23055637>). Work around this by having the test phases intentionally
+    # exit after they finish building the first time, then run the tests for real.
+    ( REALM_EXIT_AFTER_BUILDING_TESTS=YES xc "$1" ) || true
+    xc "$1"
 }
 
 clean_retrieve() {
@@ -215,12 +221,11 @@ test_ios_devices() {
         fi
         exit 1
     fi
-    cmd="$1"
-    scheme="$2"
-    configuration="$3"
+    scheme="$1"
+    configuration="$2"
     failed=0
     for device in "${serial_numbers[@]}"; do
-        $cmd "-scheme '$2' -configuration $configuration -destination 'id=$device' test" || failed=1
+        xc "-scheme '$scheme' -configuration $configuration -destination 'id=$device' test" || failed=1
     done
     return $failed
 }
@@ -405,33 +410,35 @@ case "$COMMAND" in
         ;;
 
     "ios-static")
-        build_ios_combined iOS Realm
+        build_ios_combined 'Realm iOS static' Realm "-static"
         exit 0
         ;;
 
     "ios-dynamic")
-        build_ios_combined "iOS Dynamic" Realm "-dynamic"
+        build_ios_combined "Realm" Realm
         exit 0
         ;;
 
     "ios-swift")
+        sh build.sh ios-dynamic
         build_ios_combined RealmSwift RealmSwift '' "/swift-$REALM_SWIFT_VERSION"
-        cp -R build/ios-dynamic/Realm.framework build/ios/swift-$REALM_SWIFT_VERSION
+        cp -R build/ios/Realm.framework build/ios/swift-$REALM_SWIFT_VERSION
         exit 0
         ;;
 
     "watchos")
-        build_watchos_combined "watchOS" Realm
+        build_watchos_combined Realm Realm
         exit 0
         ;;
 
     "watchos-swift")
+        sh build.sh watchos
         build_watchos_combined RealmSwift RealmSwift
         exit 0
         ;;
 
     "osx")
-        xcrealm "-scheme OSX -configuration $CONFIGURATION"
+        xc "-scheme Realm -configuration $CONFIGURATION"
         rm -rf build/osx
         mkdir build/osx
         cp -R build/DerivedData/Realm/Build/Products/$CONFIGURATION/Realm.framework build/osx
@@ -439,9 +446,10 @@ case "$COMMAND" in
         ;;
 
     "osx-swift")
-        xcrealmswift "-scheme 'RealmSwift' -configuration $CONFIGURATION build"
+        sh build.sh osx
+        xc "-scheme 'RealmSwift' -configuration $CONFIGURATION build"
         destination="build/osx/swift-$REALM_SWIFT_VERSION"
-        clean_retrieve "build/DerivedData/RealmSwift/Build/Products/$CONFIGURATION/RealmSwift.framework" "$destination" "RealmSwift.framework"
+        clean_retrieve "build/DerivedData/Realm/Build/Products/$CONFIGURATION/RealmSwift.framework" "$destination" "RealmSwift.framework"
         cp -R build/osx/Realm.framework "$destination"
         exit 0
         ;;
@@ -470,30 +478,30 @@ case "$COMMAND" in
         ;;
 
     "test-ios-static")
-        xcrealm "-scheme iOS -configuration $CONFIGURATION -sdk iphonesimulator -destination 'name=iPhone 6' build test"
+        xc_work_around_rdar_23055637 "-scheme 'Realm iOS static' -configuration $CONFIGURATION -sdk iphonesimulator -destination 'name=iPhone 6' test"
         shutdown_simulators
-        xcrealm "-scheme iOS -configuration $CONFIGURATION -sdk iphonesimulator -destination 'name=iPhone 4s' build test"
+        xc_work_around_rdar_23055637 "-scheme 'Realm iOS static' -configuration $CONFIGURATION -sdk iphonesimulator -destination 'name=iPhone 4s' test"
         exit 0
         ;;
 
     "test-ios7-static")
-        xcrealm "-scheme iOS -configuration $CONFIGURATION -sdk iphonesimulator -destination 'name=iPhone 5S,OS=7.1' build test"
+        xc_work_around_rdar_23055637 "-scheme 'Realm iOS static' -configuration $CONFIGURATION -sdk iphonesimulator -destination 'name=iPhone 5S,OS=7.1' test"
         shutdown_simulators
-        xcrealm "-scheme iOS -configuration $CONFIGURATION -sdk iphonesimulator -destination 'name=iPhone 4s,OS=7.1' build test"
+        xc_work_around_rdar_23055637 "-scheme 'Realm iOS static' -configuration $CONFIGURATION -sdk iphonesimulator -destination 'name=iPhone 4s,OS=7.1' test"
         exit 0
         ;;
 
     "test-ios-dynamic")
-        xcrealm "-scheme 'iOS Dynamic' -configuration $CONFIGURATION -sdk iphonesimulator -destination 'name=iPhone 6' build test"
+        xc_work_around_rdar_23055637 "-scheme Realm -configuration $CONFIGURATION -sdk iphonesimulator -destination 'name=iPhone 6' test"
         shutdown_simulators
-        xcrealm "-scheme 'iOS Dynamic' -configuration $CONFIGURATION -sdk iphonesimulator -destination 'name=iPhone 4s' build test"
+        xc_work_around_rdar_23055637 "-scheme Realm -configuration $CONFIGURATION -sdk iphonesimulator -destination 'name=iPhone 4s' test"
         exit 0
         ;;
 
     "test-ios-swift")
-        xcrealmswift "-scheme RealmSwift -configuration $CONFIGURATION -sdk iphonesimulator -destination 'name=iPhone 6' build test"
+        xc "-scheme RealmSwift -configuration $CONFIGURATION -sdk iphonesimulator -destination 'name=iPhone 6' build test"
         shutdown_simulators
-        xcrealmswift "-scheme RealmSwift -configuration $CONFIGURATION -sdk iphonesimulator -destination 'name=iPhone 4s' build test"
+        xc "-scheme RealmSwift -configuration $CONFIGURATION -sdk iphonesimulator -destination 'name=iPhone 4s' build test"
         exit 0
         ;;
 
@@ -506,12 +514,12 @@ case "$COMMAND" in
         ;;
 
     "test-ios-devices-objc")
-        test_ios_devices xcrealm "iOS Device Tests" "$CONFIGURATION"
+        test_ios_devices "Realm iOS static" "$CONFIGURATION"
         exit $?
         ;;
 
     "test-ios-devices-swift")
-        test_ios_devices xcrealmswift "RealmSwift" "$CONFIGURATION"
+        test_ios_devices "RealmSwift" "$CONFIGURATION"
         exit $?
         ;;
 
@@ -520,12 +528,12 @@ case "$COMMAND" in
         if [[ "$CONFIGURATION" == "Debug" ]]; then
             COVERAGE_PARAMS="GCC_GENERATE_TEST_COVERAGE_FILES=YES GCC_INSTRUMENT_PROGRAM_FLOW_ARCS=YES"
         fi
-        xcrealm "-scheme OSX -configuration $CONFIGURATION test $COVERAGE_PARAMS"
+        xc "-scheme Realm -configuration $CONFIGURATION test $COVERAGE_PARAMS"
         exit 0
         ;;
 
     "test-osx-swift")
-        xcrealmswift "-scheme RealmSwift -configuration $CONFIGURATION test"
+        xc "-scheme RealmSwift -configuration $CONFIGURATION test"
         exit 0
         ;;
 
@@ -789,29 +797,29 @@ case "$COMMAND" in
         REALM_SWIFT_VERSION=1.2 sh build.sh prelaunch-simulator
         REALM_SWIFT_VERSION=1.2 sh build.sh test-ios-static
         REALM_SWIFT_VERSION=1.2 sh build.sh ios-static
-        move_to_clean_dir build/ios/Realm.framework xcode-6
+        move_to_clean_dir build/ios-static/Realm.framework xcode-6
         rm -rf build
 
         REALM_SWIFT_VERSION=2.1 sh build.sh prelaunch-simulator
         REALM_SWIFT_VERSION=2.1 sh build.sh test-ios-static
         REALM_SWIFT_VERSION=2.1 sh build.sh ios-static
-        move_to_clean_dir build/ios/Realm.framework xcode-7
+        move_to_clean_dir build/ios-static/Realm.framework xcode-7
 
-        zip --symlinks -r build/ios/realm-framework-ios.zip xcode-6 xcode-7
+        zip --symlinks -r build/ios-static/realm-framework-ios.zip xcode-6 xcode-7
         ;;
 
     "package-ios-dynamic")
         cd tightdb_objc
         REALM_SWIFT_VERSION=1.2 sh build.sh prelaunch-simulator
         REALM_SWIFT_VERSION=1.2 sh build.sh ios-dynamic
-        move_to_clean_dir build/ios-dynamic/Realm.framework xcode-6
+        move_to_clean_dir build/ios/Realm.framework xcode-6
         rm -rf build
 
         REALM_SWIFT_VERSION=2.1 sh build.sh prelaunch-simulator
         REALM_SWIFT_VERSION=2.1 sh build.sh ios-dynamic
-        move_to_clean_dir build/ios-dynamic/Realm.framework xcode-7
+        move_to_clean_dir build/ios/Realm.framework xcode-7
 
-        zip --symlinks -r build/ios-dynamic/realm-dynamic-framework-ios.zip xcode-6 xcode-7
+        zip --symlinks -r build/ios/realm-dynamic-framework-ios.zip xcode-6 xcode-7
         ;;
 
     "package-osx")
@@ -825,7 +833,7 @@ case "$COMMAND" in
     "package-ios-swift")
         cd tightdb_objc
         for version in 1.2 2.1; do
-            rm -rf build/ios-dynamic
+            rm -rf build/ios/Realm.framework
             REALM_SWIFT_VERSION=$version sh build.sh prelaunch-simulator
             REALM_SWIFT_VERSION=$version sh build.sh ios-swift
         done
@@ -971,11 +979,11 @@ EOF
 
         echo 'Packaging iOS static'
         sh tightdb_objc/build.sh package-ios-static
-        cp tightdb_objc/build/ios/realm-framework-ios.zip .
+        cp tightdb_objc/build/ios-static/realm-framework-ios.zip .
 
         echo 'Packaging iOS dynamic'
         sh tightdb_objc/build.sh package-ios-dynamic
-        cp tightdb_objc/build/ios-dynamic/realm-dynamic-framework-ios.zip .
+        cp tightdb_objc/build/ios/realm-dynamic-framework-ios.zip .
 
         echo 'Packaging OS X'
         sh tightdb_objc/build.sh package-osx